Фикс зависания на машинах, не поддерживающих FrameBuffer. Правки шейдеров. Корректировка заготовки теней.

+++ shaders/FontColorizerShader.frag
@@ -16,9 +16,9 @@ vec3 Color16To32(int c)
 
 	vec3 v;
 
-	v.r = AND32(c / 1024.0);
-	v.g = AND32(c / 32.0);
-	v.b = AND32(c);
+	v.r = float(AND32(c / 1024));
+	v.g = float(AND32(c / 32));
+	v.b = float(AND32(c));
 
 	return (v / mod32);
 }
@@ -31,7 +31,7 @@ void main(void)
 	{
 		if (drawMode == 1 || (drawMode == 2 && textureColor.r == textureColor.g && textureColor.r == textureColor.b))
 		{
-			int index = AND32(textureColor.r * 31.875);
+			int index = AND32(int(textureColor.r * 31.875));
 
 			vec3 resultColor = Color16To32(colors[index]);
